Dashboard API: relocate /:team/:cluster to /:cluster/:team
Affected Pages:
1. Home
2. Clusters
3. Cluster
